fix: Created a partial class per platform
